refactor: rebrand ProposalsDetails
