+++ packages/desktop/components/index.ts
@@ -17,7 +17,7 @@ export { default as NetworkCard } from './NetworkCard.svelte'
 export { default as NodeListTable } from './NodeListTable.svelte'
 export { default as ProposalAnswer } from './ProposalAnswer.svelte'
 export { default as ProposalCard } from './ProposalCard.svelte'
-export { default as ProposalsDetails } from './ProposalsDetails.svelte'
+export { default as ProposalListDetails } from './ProposalListDetails.svelte'
 export { default as SidebarTab } from './SidebarTab.svelte'
 export { default as StatusTile, type StatusTileProps } from './StatusTile.svelte'
 export { default as TitleBar } from './TitleBar.svelte'
