+++ packages/shared/src/lib/contexts/governance/interfaces/index.ts
@@ -2,4 +2,4 @@ export * from './organization.interface'
 export * from './proposal-answer-percentages.interface'
 export * from './proposal-filter.interface'
 export * from './proposal-state.interface'
-export * from './proposals-details.interface'
+export * from './proposal-list-details.interface'
